requirements.txt
Django==1.8.7
argparse==1.2.1
wsgiref==0.1.2
Pillow==3.1.0
django-cors-headers==1.1.0
pytz==2015.7
South==1.0.2
factory-boy==2.1.1
fake-factory==0.5.3

dj-database-url==0.4.0
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6

